test/requirement.txt
sentence-transformers
faiss-cpu
torch
transformers
langchain
accelerate
sounddevice
numpy
threading2
keyboard
scipy
soundfile
datasets
langchain-community
bitsandbytes
